+++ study/GCD.cpp
@@ -16,7 +16,7 @@ using ull = unsigned long long;
 using Pii = pair<int, int>;
 using Pll = pair<ll, ll>;
  
-int GCD(int a, int b) {
+ll GCD(ll a, ll b) {
     if (a <= 0 || b <= 0) {
         return -1;
     }
@@ -39,8 +39,8 @@ int GCD(int a, int b) {
 
 int main(int argc, char *argv[])
 {
-    int a = 2;
-    int b = 3;
+    ll a = 2;
+    ll b = 3;
     cout << GCD(a, b) << endl;
     return 0;
 }
