+++ study/algorithm.cpp
@@ -18,10 +18,10 @@ using Pll = pair<ll, ll>;
  
 int main(int argc, char *argv[])
 {
-    vector<int> a, dummy;
+    vector<ll> a, dummy;
     
-    int N = 10;
-    for (int i = 0; i < N; i++) {
+    ll N = 10;
+    for (ll i = 0; i < N; i++) {
         a.push_back(i);
     }
 
@@ -37,16 +37,15 @@ int main(int argc, char *argv[])
     //copy(a, a + N, dummy.begin());
     copy(a.begin(), a.end(), back_inserter(dummy));
 
-    for (int i = 0; i < a.size(); i++) {
+    for (ll i = 0; i < a.size(); i++) {
         cout << a[i] << endl;
     }
 
-    for (int i = 0; i < dummy.size(); i++) {
+    for (ll i = 0; i < dummy.size(); i++) {
         cout << dummy[i] << endl;
     }
 
     //fill(a, a + N, 0);
     fill(a.begin(), a.end(), 0);
-
     return 0;
 }
